Change Contexter to Requester for more options
When doing any kind of log/context propagation across services there
will be a need to convey operational details, such as retry count,
across the HTTP boundary. To support this, the Contexter interface
is replaced with a Requester interface in order to give retry policies
full access to the request object before it is executed.