Update default transport to use clone feature
This helps ensure the OptionDefaultTransport is always up to date by
leveraging the built-in Clone() method of the *http.Transport.

The added test is a basic check to ensure that all the comparable values
match. Additional fields may appear over time but this _should_ be
enough to prove that the Clone() feature is working as expected.