Fix hedger done context usage
This fixes an issue where the hedger plugin will prematurely cancel a
request context for a successful call. The standard library
http.Transport wraps the response body in a context aware reader that
causes calls to resp.Body.Read() to fail even if the response was a
success and there is a body to read. This patch adjusts the context
management so that "terminal" response returned by the hedger will
always have an active context.