Use deep copy when cloning requests
The previous version of this code uses a shallow copy of the request
object when retrying or hedging. So long as shared attributes of the
request, such as the headers map, is read-only during the retry/hedge
then there is no issue. However, if the retry or hedger is wrapped
around subsequent middleware that modifies a shared asset, such as
injecting a header value, then it creates a race condition and panics.

I've added a test case that reliably triggers the race detector during
tests and switched the shallow copy to a deep copy of the request.